Delete application action; move standalone action into main class
